update Tests classes
