+++ core/src/test/java/test/failedreporter/issue2517/FailedDataProviderWithFactoryReporterTest.java
@@ -1,0 +1,59 @@
+package test.failedreporter.issue2517;
+
+import org.testng.Assert;
+import org.testng.TestNG;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+import org.testng.collections.Lists;
+import test.SimpleBaseTest;
+
+import java.io.File;
+import java.util.List;
+
+public class FailedDataProviderWithFactoryReporterTest extends SimpleBaseTest {
+    private File mTempDirectory;
+
+    @BeforeMethod
+    public void setUp() {
+        mTempDirectory = createDirInTempDir("testng-tmp-" + System.currentTimeMillis() % 1000);
+    }
+
+    @AfterMethod
+    public void tearDown() {
+        deleteDir(mTempDirectory);
+    }
+
+    @Test
+    public void failedMethodWithDataProviderAndFactoryShouldHaveInvocationNumbers() {
+        testFailedReporter(
+                new String[]{"f1"},
+                "<include name=\"%s\" invocation-numbers=\"1\"/>",
+                FailedDataProviderWithFactoryReporterSampleTest.class);
+    }
+
+    private void testFailedReporter(String[] expectedMethods, String expectedLine, Class<?>... cls) {
+        triggerTest(cls);
+        runAssertions(mTempDirectory, expectedMethods, expectedLine);
+    }
+
+    private void triggerTest(Class<?>... cls) {
+        TestNG tng = create(mTempDirectory.toPath(), cls);
+        tng.setUseDefaultListeners(true);
+        tng.run();
+    }
+
+    static void runAssertions(File outputDir, String[] expectedMethods, String expectedLine) {
+        runAssertions(outputDir, expectedMethods, expectedLine, 1);
+    }
+
+    private static void runAssertions(
+            File outputDir, String[] expectedMethods, String expectedLine, int expected) {
+        File failed = new File(outputDir, "testng-failed.xml");
+        for (String s : expectedMethods) {
+            List<String> resultLines = Lists.newArrayList();
+            grep(failed, String.format(expectedLine, s), resultLines);
+            Assert.assertEquals(resultLines.size(), expected);
+        }
+    }
+}
